requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1